Added recursive func that finds paths to target.
Now i have to find out how to materialize these paths to compare each of them and output the best path possible.
I was thinking about another recursive call that makes every success node the target position, so the path gets smaller and smaller to the point that the initial position has the target position as a possible move.
The difficult part is to find a way to organize these paths into an array (or another data structure.)