Finished project + optimized path.
Because the algorithm was creating having in mind the function going from a lower initialPos to a bigger finalPos, when the initialPos is bigger than the finalPos, the path wasn't always the best possible.
So after doing many tests, the initialPos (smaller) to finalPos (bigger) is always shorter than the other way around.
What i did is, when we call a big initialPos, we just swap initialPos to finalPos, but we need to keep the logic, so the path start to end, will be backwards too, showing the best logical path.